// client/src/types/searchFlight.type.ts
export interface IBaggages {
  hand: number;
  hold: number;
}

export interface ISearchFlightRequest {
  fly_from: string;
  fly_to?: string;
  date_from: string;
  date_to: string;
  return_from?: string;
  return_to?: string;
  nights_in_dst_from?: number;
  nights_in_dst_to?: number;
  max_fly_duration?: number;
  ret_from_diff_city?: boolean;
  ret_to_diff_city?: boolean;
  one_for_city?: number;
  one_per_date?: number;
  adults?: number;
  children?: number;
  infants?: number;
  selected_cabins?: string;
  mix_with_cabins?: string;
  adult_hold_bag?: string;
  adult_hand_bag?: string;
  child_hold_bag?: string;
  child_hand_bag?: string;
  fly_days?: string;
  fly_days_type?: string;
  ret_fly_days?: string;
  ret_fly_days_type?: string;
  only_working_days?: boolean;
  only_weekends?: boolean;
  partner_market?: string;
  curr?: string;
  locale?: string;
  price_from?: number;
  price_to?: number;
  dtime_from?: string;
  dtime_to?: string;
  atime_from?: string;
  atime_to?: string;
  ret_dtime_from?: string;
  ret_dtime_to?: string;
  ret_atime_from?: string;
  ret_atime_to?: string;
  stopover_from?: string;
  stopover_to?: string;
  max_stopovers?: number;
  max_sector_stopovers?: number;
  conn_on_diff_airport?: number;
  ret_from_diff_airport?: number;
  ret_to_diff_airport?: number;
  select_airlines?: string;
  select_airlines_exclude?: boolean;
  select_stop_airport?: string;
  select_stop_airport_exclude?: boolean;
  vehicle_type?: string;
  enable_vi?: boolean;
  sort?: string;
  limit?: number;
}